[Refactor] Share offline profiler argument parser
